P3.2: Texture Compression Strategies (partial, stable)
This integration test loads texture assets in different formats. The
goal is to measure which format is heavier on disk, which one takes
the least to load in memory, which one occupies the least RAM once
decompressed.

This partial implementation contains the architectural blueprint of
the solution and a few texture formats, to get some early results.

For now, JPG loading & rendering is fully implemented, plus the
boilerplate code for the all other formats.
ASTC is also ready. ETC2 in progress.
All the others formats are pending.

I refactored a lot of the existing code to maximize reusability when
working with textures or loading assets.

I also got a zsh script that compresses entire directories of PNG &
JPG files in order to create ASTC, ETC2 or BASIS counterparts.

I expanded the install_deps script to include basis_universal.

Test: I tested this local and manually.

Change-Id: I5851ffb7692848471b35a7a3759e5202f3f29e9d
Signed-off-by: Diego Dagum <dagum@google.com>