P3.2: Texture Compression Strategies (WIP)
This integration test loads texture assets in different formats. The
goal is to measure which format is heavier on disk, which one takes
the least to load in memory, which one occupies the least RAM once
decompressed.

For now, JPG loading & rendering is fully implemented, plus the
boilerplate code for the all other formats.
ASTC is in progress, and all the others to follow.

I also got a zsh script that compresses entire directories of PNG &
JPG files in order to create ASTC, ETC2 or BASIS counterparts.

Test: I tested this locally.

Change-Id: I5851ffb7692848471b35a7a3759e5202f3f29e9d
Signed-off-by: Diego Dagum <dagum@google.com>