Update patch set 6
Patch Set 6: Presubmit-Ready+1

This is fixing the AAR generation so that we can compile a samples with it.
Shared libraries are for now excluded from the AAR while I understand if the error that Prefab output when they are included is an issue in the AAR or in prefab.

Patch-set: 6
Reviewer: Gerrit User 1373878 <1373878@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1554536 <1554536@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1498014 <1498014@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1