Implement yellow warning states for some heuristics.
Stresser tests are prematurely resuming allocation after a warning for some heuristics where the metric tested is 'hovering' between red and green. This can be interpreted by the test that the warning state is over and allocation can be resumed. In reality the benefits of the memory release haven't taken effect.

Implementing the 'yellow' state for those heuristics will fix this for the most part because a decent sized band will make 'hovering' into the green (e.g. going green for some reason other than effective clearing of resources) much less likely.

Bug: 156223565
Test: Manually
Change-Id: I277f13245a60d8846424eb3a7d7c6426ce9496b6