Update patch set 16
Patch Set 16: Code-Review+1

This is a VERY tough refactoring, and it's too big for me to be confident in just by looking at it. But given the test output, quality of code, and the thoroughness of documentation, I am inclined to trust it.

Patch-set: 16
Reviewer: Gerrit User 1557643 <1557643@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1