Add gaming-performance-tuner AAR and fix fullsdk build
Add a target to build the AAR for Tuning Fork.
Add a target to build the AAR for Swappy inside fullsdk (duplication to
be removed later).
Fix building of Tuning Fork samples (wrong ndk version for build bots)

Test: DIST_DIR=/tmp/test ./build.sh full
Test: verify that AARs are built, gamesdk.zip is built, that it contains
the 4 sample apks, TuningForkMonitor apk, the libs for tuningfork,
unitytuningfork, swappy, the static lib (libgamesdk.a)

Test: build the AARs, copied them to a separate folder, used them to
build scaledballs (insightsdemo) and bouncyball.

Bug: 157563056
Bug: 147287876

Change-Id: I724c45debdf9d97f7d8845f93b640091ff1aec1c