Restructure Score class a little.
Break out two new methods, writeReport and writeTable.

Two reasons. 1) to manage the complexity a little, the main methods is getting huge. 2) prepare for future improvements where a) the report is written as each batch of results is received so that you can view the report without waiting for every run to finish and b) multiple tables will be rendered groupling the data in different ways.

Test: Manual
Change-Id: Ie82fa0873334c48d91b24e284410787870baa3da