checkHeuristics now takes the metrics dictionary made in getMemoryMetrics().
This makes getMemoryMetrics() the only place metrics are extracted and make it a source of truth for metrics.

This will eventually form how clients get metrics from the API.

The Info class now exposes the 'baseline' (first set of metrics - this will be used in some heuristics later).

One format change is that metrics are now in their own dictionary with key 'metrics'.

Bug: 150696984
Test: Manual
Change-Id: I8fa6d7d45127349ca3936abcc99eed38ea03f6c1