Report memory warnings as a JSONArray, not callbacks.
The next generation of heuristics allow for multiple heuristics with the
same 'name' but different parameters.

I will need to see in the results log which one fired.

Presently we just include the name but that won't allow me to
distinguish which one it was. We should include the full heuristic spec
in the report.

Rather than converting the callbacks into JSON for the report, it's
more straightforward to form the results in JSON in the first place.

IIRC bkaya had concerns about callbacks rather than returning a
container in the last CL that changed the interface.

Other benefits of the approach-
- Trusted testers can use the warning JSON in C# code without having to
write a wrapper around the callback system.
- Trusted testers can more easily record the warnings we sent them in
telemetry (it's just a JSON string now).
- I can add memory estimates in the same object (which is why I've
wrapped the warnings in 'warnings' object).

'Green' signals are not sent. Absence of a signal is the same as a green
signal. This avoids clutter in the result JSON.

We might reconsider this down the line when the library is more tested,
and it is not worth the slightly unusual step of using JSON containers
in an internal library.

Bug: 159105919
Test: Manual
Change-Id: I167250ef76d092b3f1e8e40fc56058b1507db9c2