Write interim versions of the report as results are being collected.
This make a big impact on the usability of the tests, as in large tests it is common for most results to be in after 15 minutes, but if you wait until they are all in you could be waiting an hour or more.

The report is written whenever there are new reults to include but no new result have been received for 10 seconds.

Bug: 159204344
Test: Manual
Change-Id: I095634087fba7840c1094af06e745add9d84d54c