Unify all generated test allocated values and move them from metrics.
They're not metrics per-se, they are generated by the app.

It's getting a bit out of hand having six of these values. Plus it's going to make it harder to have combined tests, if I do that. So I put them all in a single dictionary and sum them in the report generators.

Bug: 159211848
Test: Manual
Change-Id: I9a804de4529b89cfe9965a9a827ce1b574a80cf0