First version of live memory prediction.
The device settings now includes a full set of 'baseline' and 'limit' metrics, plus the calculation of how much memory was allocated ('applicationAllocated'). (Discover tool updated to generate this)

These are used to perform the calculation.

Predictions are renderer on the graph as a total of memory allocated so far and the prediction - that way a good prediction should appear as a near horizontal line.

Bug: 145291358
Test: Manual
Change-Id: Ida743fa23fb372330cb3dd97396afb29769ddf36