Fix a problem with the code that generates the key for results.
We need to find a key that individually identifies lab devices from each result, so we can group a device's results together on their own when we receive them. This is harder than it sounds are there can be multiple devices with the same device ID. We can use the device data ('build') returned by the device but I have seen occasions where two similar devices have the same build data. Using Firebase's own device values (the 'device dimensions') is more reliable.

I used to generate a dictionary of device dimensions in the Collector (and copy it into 'extra') but at some point I took that out and I forgot to update this code.

We get all the step data from the Firebase API now in 'extra' so I can access it directly here.

Test: Manual
Change-Id: I49d01774b85e769076c91068bd72255a6db0464e