Use a folder based on the historyId of the run if possible.
historyId is the Firebase identifier for a run. It is only known for sure (the way the code is arranged at the moment) when the first result comes in.

If we use this value to make the results folder we can reuse the old folder when extra reports are generated from the same run (e.g. we changed the report generation code and wanted to run it again). This is neater. Plus it's much easier to see the historyId for previous runs, which makes it easier to rerun the report generator for particular old runs.

Test: Manual
Change-Id: Ib3f01b072b8a4b271121b305ce7734525398dba9