Force set one scenario in the launcher.
iStresser launcher can be used to launch any helperlibrary instrumented app (this is what we did with Wildlife).

I basically override the scenario system and always set scenario 1, but upload custom parameters in a JSON file.

Firebase game loops requires this code in AndroidManifest.xml

<meta-data android:name="com.google.test.loops" android:value="1" />

If the implementer  gets the number of test loops wrong (e.g. they're copy pasting from the sample code which has it set at something else IIRC) then the run gets multiplied by that many times.

As a precaution we can simply force the number of scenarios to be 1.

Bug: 159422095
Test: Manual
Change-Id: I4d11c51a5689339936104f173f11ca05b5606381