The dictionary merger in the report page uses the same method as the app and launcher.
I already reworked the code that combines params from multiple dimensions ('flattening') so that dictionaries were deeply merged and arrays concatenated. (See CL linked on the bug)

There's a similar 'flatten' that happens on the report page so we can show what the actual run params were on the launch.

This CL updates it to use the same method as the Java implementation.

Also remove a 'debugger' statement committed in error.

Test: Manual
Bug: 157280190
Change-Id: Icb42ad0f2bc93bd82e6bc8afb91e55b579a4ed24