Series are always colored the same across graphs, based on their names.
Since the selection of series can vary between different runs, the colors selected by dygraph can vary too.

It makes it easier to interpret graphs if the colors are consistent.

This CL applies a hue shade based on the hash code of the metric name.

Series that are used in active heuristics are drawn more thickly to emphasize them.

It also removes the assumption that applicationAllocated appears in the metrics (not the case for some runs).

Test: Manual
Change-Id: Ic8326964e6ea547ac3d30ace335563c8e98c69dd