Change how results are gathered.
When gathering results store them in a map keyed on the coordinates, rather than in a list ordered by the flattened order of the coordinates.

This is preparing for a change where the whole table will be output multiple times in different grouping orders (there won't be one definitive flattened order to use).

Bug: 156226431
Test: Manual
Change-Id: I8ae6de1e958552e8272d117ef954cc31d48c6ddc