Don't launch the completed reports in the browser automatically.
Instead of forcibly launching the report in the browser, log the report URLs as they are written to the console, so that I can chose to click on them and launch them manually.

It's much more common to view reports as they are being generated now. The auto launch can be annoying.

Test: Manual

Change-Id: I9562466139a4d9f92b4f53482aee3ccdaaa8fd30