Simplify API for the client.
The idea is to reduce the information the client has to supply when it calls the API, by encapsulating the information within the API.

I had Heuristic inherit from Info to get access to the data it needs. deviceSettings was moved to that class because it can be.

I am going to rename both classes soon to better reflect what they do.

Bug: 145291358
Test: Manual
Change-Id: I416567f6beffd5939796f577fbf59b08bf642df9