Simplify the API; clients don't have to fetch metrics, metrics loggable from 'advice'
In iStresser, all of 'advice' is now recorded in the metric logs.

This reflects what I want trusted testers to do (it's easier than giving them complex instructions about what to log).

Metrics is now part of this. This also simplifies the requirements to get metrics before passing it in to the advisor - the advisor can get it.

We don't need the logic about including 'predictions' and 'warnings' selectively - they always go in as part of 'advice'.

I no longer create the warnings array if there are no warnings. This reduces the log spam.

We no longer need the time measurement (under 'meta') in getMemoryMetris as it's covered by the getAdvice timing.

Since metrics is included under advice we should take it out of standardInfo so as not to take two sets of metrics. That means metrics needs putting back in in the places where advice is not taken.

For metrics on each row, the report generator will look under 'advice/metrics' in the first instance. Then it looks under 'metrics'.

Bug: 145291358
Test: Manual
Change-Id: Ifa5bf69212c5a5a850738bd7b064f7a6504a6c2a