Group the information we want from trusted testers getDeviceInfo()
The point is to make it as easy as possible to get the information we want in telemetry, debugging etc. without having to give complex instructions to devs. We just ask for this object.

This includes the baseline, the device data, the device profile selected and the parameters used.

This also reduces the number of public methods in the API.

Since the paramters are the 'flat' version we will have both the flat and dimension array form so we don't need to reflatten the parameters when generating the report.

Bug: 145291358
Test: Manual
Change-Id: I9c277932a955fb18cbab3e5a15830c352ac6a035