Change how time is stored (again).
This puts time storage back into metrics collection. The reasoning is this: if we get telemetry back from clients we will need to know the time the samples were taken to graph the results. So I want the collection to be automatic.

I also want the grapher to be able to handle these types of result, so we can use it for trusted tester reports. (I did this once with Wildlife already)

Therefore iStresser's time storage should come out (also prevents duplication).

This does mean that time is stored with the metrics, which can be in different places. So it does complicate the code that has to interpret the results. So, in short, we are complicating the interpreters so that the API can be simpler for devs to use.

There's also some overdue fixes to the Discover utility in there.

Bug: 145291358
Test: Manual
Change-Id: Ic4c72704cb636189af8f03c35accc2175b1f715b