Score reports use a new version of the data explorer
The data explorer is a function I created for the reports to allow semi-structured data to be presented to the user in a table that can conceal details until expanded. I want to use the same system to present the parameters in the score table, because they can now get highly complex.

Because of the way the report generator works that version runs in the report's JavaScript. The Score report is built server side, so needs a Java version of the same thing written (we should prefer to do things on the server in most cases anyway).

I want to now build the server-side HTML using structured w3c classes similar to the JS. Currently the Java side rendering is done with text concatenation, but that is error-prone (easy to break the whole structure by dropping a closing tag, for example). I've added a helper class to allow w3c classes to build HTML, and I intend to convert the existing code at some point.

Test: Manual
Change-Id: Iaef03763c2805028ee0fcc1f7cacf5c4603459fe