Metrics are grouped into their own dictionaries.
This makes the output and lookup table far easier to read, in turn making it easier to manually configure which metrics are collected.

Since I had to rework how the predictions code gets the metrics, I also made the predictions configuration set in the params.

I generated a new lookup table for lab devices which includes all parameters.

Bug: 161580690
Test: Manual
Change-Id: I0bfb3aaa5163fe685e008de153697fa60a281a2c