Refactor builds of samples into the Gradle build
This removes most of the logic inside the build.sh, instead using a project flag (-PincludeSampleArtifacts) to decide if the samples must be built and packaged in the zip.
This means that packageZip is now able to generate the "complete" archive, ready to be distributed, and that there is no additional work done by a script.

When adding a new library, samples can then be declared in the build.gradle and will be packaged as declared with no extra work involved to change the build.sh.

Also removes the DIST_DIR environment variable usage and clean a bit the build tasks.

Bug: 161773980
Test: ./gradlew packageZip -Plibraries=swappy,tuningfork -PincludeSampleSources -PincludeSampleArtifacts
Test: presubmit generated gamesdk.zip content manually verified for gamesdk-samples. Content is same as previously.
Change-Id: I7946febb82374f17d1ff755a1f352403ead736e1