Create on device stress tester.
The on device stress tester is a working (but experimental) class to stress test the device in a separate process.

It is added to the MemoryAdvisor when onDeviceStressTest dictionary is found in the params. Data from the tester is used as baseline and limit for the duration of the test.

Because the MemoryAdvisor needs these values, and they take 5 to 30 seconds to generate, an asynchronous interface is needed (the "ready handler"). For compatibility this is optional, if onDeviceStressTest isn't needed.

iStresser needed to be reengineered with this in mind. It meant that the deviceInfo cannot appear in the first row of the output, because it may not be known for a while after startup. All the testers therefore needed to be adapted to find the deviceInfo wherever it is.

The library now needs a method to occupy memory with non-compressible data; sadly this can't be shared with iStresser because native->native calls are not officially supported in Android libraries. We'd have to go out to Java.

The library now requires iStresser's getOrDefault() method.

Bug: 158520924
Test: Manual
Change-Id: If742d75ff2f4ccd4694a1c7688765415db632d0d