Added endless-tunnel ndk sample to the third-party/ directory
Copied the contents of https://github.com/android/ndk-samples/tree/master/endless-tunnel
to the third-party/ directory, and added the licensing information
present on the ndk-samples.

The only change I added was to revert the gradle version back to 5.4.1,
and include a gradle link to this project from samples/tuningfork/endless-tunnel.

Test: Made sure the app runs from the samples/tuningfork/endless-tunnel folder
Change-Id: Ia18b71534ab46623b6852987bf47fe1844bd3b97