Convert collate from Maven to Gradle build.
First reason is that the other parts use Gradle and it was confusing to keep switching between the systems; particularly when engineering new dependencies in common.

Second reason is that in my experience Gradle is easier to get running from a fresh repo. Maven seems a lot more brittle for some reason.

Test: Manual
Bug: 163192107
Change-Id: Iaa2ca2117c56c4355c3bc10c4694a2a26419e1e7