iStresser uses MemoryWatcher instead of its own timer.
This is to ensure it's better tested.

I had to adapt MemoryWatcher so that it can provide the full memory advice object, which in turn led to a change to the interface so that clients don't *have* to provide a listener for this method.

Version also increased (of both libraries - to avoid confusion).

Also included an overdue fix for the new andoridx requirement, and an update to the docs about androidx. Plus an overdue fix for a weird version of gradle on the example project.

Bug: 170370813
Test: Manual
Change-Id: I577642dcc73dc324ca9a822259ca5d6f0f189efe