Fix combo box table UI bug
There was a bug that was noticed only in combo box which happened due to focus problem. The combo box would receive focus call twice which leads it to opening up the popup then closing it instantly. and the user will have to open it up again. which was a bad user experience. This CL fixes this bug by setting the comboBox to be non-focusable so only one focus call is called and drawing border manually.

TableComboBox class is designed to be used only for tables. as this problem occurs only in tables. otherwise JComboBox or ComboBox is usually a good fit.

Test: Manual Testing of the plugin using experimentsdemo

Change-Id: Id4e017a3e6286d80623796cfd43c142a930e304a