Create a callback interface for the on-device stress test.
This is being used by the standalone iStresser (to be called User Benchmark). It's much nicer to be able to report progress to the user.

It's conceivable that other applications might want to report progress if they were to use this feature.

It's undocumented at the moment but spinning on the version number anyway because the API is changed.

Bug: 168468341
Test: Manual
Change-Id: Ic8f18abfed96f0b99ff53d24f75858651b81c663