Fix add_gamesdk_target compiling more libraries than asked
If the CMake cache was containing options to build libraries, these options would be persisted and reused when compiling different samples. This is usually fine, but also means compiling useless libraries for some samples.

This change ensures that only the specified libraries are compiled, no more, by using a single variables to specify the libraries to build.

Bug: 182893746
Change-Id: I4cfc9cc78f23e971ab5f315443208ecb930e6966