iStresser: Stopping the practice of storing the metrics with every row.
Two reasons: 1. it means the line can by logged with less delay, 2. to prepare for the removal of the responsibility of MemoryAdvisor to vend current metrics, itself to prepare for cleanup of the confusing overlap of responsibilities of MemoryAdvisor and MemoryMonitor. Also closer to the possiblity of iStresser not needing a MemoryAdvisor directly at all.

Test: Manual
Bug: 187912947
Change-Id: I05ddaf5f838044fdf37236601fb98e0204961754