MemoryMonitor made much closer to its original intention; a class to take current memory readings.
State is reduced; doesn't store a field selection internally any more, so the parameterless getMemoryMetrics goes.

Responsibility of recording the baseline (initial) readings, and non-memory information about the device is to the higher-level MemoryAdvisor, which is the only consumer of this data anyway.

The two ML predictors are moved to MemoryAdvice, where they makes more sense as they are generating advise, rather than gathering raw data.

Test: Manual
Bug: 187912947
Change-Id: I8cf157449b0ae2d3890047911c2454ca87cf44b5