Added SOC_MODEL and SOC_MANUFACTURER info to requestinfo.
UploadTelemetryRequest is now also populated with the system on chip
model and manufacturer info, if the device version is at least S.

Note that this feature won't work until the S version is public,
since on prerelease versions of android S, sdk version returns
30 instead of 31.

Test: Used experiments demo on an S and a non-S device to confirm
that it is working in both cases. Also ran TuningForkTestApp.

Change-Id: I4846a438ca53f08a7fc68d826fc0d0583b3044a5