Update to the latest vesion of Gradle and the Android Gradle plugin that support Java 8.
Per b/198778393 we can't currently run Java 11 on the buildbot (until someone works out how to do it).

This restricts the use of the latest Gradle.

I still want to use the same version of Gradle and the Android Gradle plugin when building my projects, because switching causes the daemon to change which displays errors about side effects.

This also may help delay that time we are forced to upgrade.

Bug: 198778393
Test: Manual
Change-Id: Ieeceac645212016e4e137536f8b20fcba943509f