TuningFork workaround fix for corrupted memory in api key for API <= 23.
For some reason, the memory location corresponding to the api key ends up containing garbage values instead of nullptr when the struct is passed from C# on devices with API <= 23. This results in failing to resort to the tuningfork_settings.bin for reading the correct api key. This workaround fix allows to correctly read the api key from the binary.

Test: tested on firebase devices.

Change-Id: Ie9fd03b9ba126b4695ec619c81473bb90cda3c25