AGDKTunnel - Add Memory Advice library support
Adds Memory Advice integration to AGDKTunnel. See the README
for implementation particulars. Summary: when active (can be turned
on by the Consume Memory button on the title screen),
a MemoryConsumer will make allocations and respond to memory advice
signals by purging allocations. When active, MemoryConsumer displays
memory status and statistics onscreen in the UI and Play scenes.

Test: Build and run in Android Studio, note behavior in onscreen
memory statistics as documented in the README.

Change-Id: I9af278e5433362869cc624ea91d7719517afa4c8