Revert "Allow derived classes of GameActivity to handle native library loading"
This reverts commit d1b19289acbd317ae5ceea55dac4d5f13cbaf235.

Reason for revert: There is a bug in this code, this CL was supposed to include an onCreate method which passes 0 to the nativeHandle by default to avoid breaking anyone depending on the old method. We are discussing how we should change the native library loading. I will revert this until we have decided how we want to handle native library loading.

Change-Id: I91c6295396a31cf632121e1b21642e80c1398079