Remove unnecessary return type
Some of the methods in text renderer are not expected to
return anything. One reason for such return type is to chain calls
to the method. However it is not done on all the methods and it is
never used that way in the sample. Change their definition to reflect that.

Bug:
Test: Ran agdktunnel sample
Change-Id: I4606074e0fdb7c702661c8e763a4c41ebbfd6d3a