Add frame-pacing-reset API to Swappy
In certain instances, the user may wish to reset the frame pacing state
and start over with a fresh state. The new API introduced would allow
the user to do that. It is implemented by setting the swap duration and
refresh period to the initial values at the end of the frame. The reset
takes effect from next frame.

Bug: b/263872091
Test: Modified Bouncyball to drop the workload and call the reset.
Change-Id: I774e420c0d69132611b8780ab8757b92b808cf01