Don't instruct to add ninja to the path for the locally installed tools case.
For that case, add a note that the required tools should be downloaded via SDK Manager (the precise tool versions are hard to predict so can't practically be listed).

Also, include Cmake as this is required to be set in the path in the same way Ninja is.

Instructions for Ninja path configuration moved to the one section where it is relevant. It should not be required for the "Build with all prebuilt SDKs" section as the whole ANDROID_HOME location is changed to one that includes ninja and cmake.

Bug: 288079718
Test: N/A
Change-Id: I0142d327cd5375ce103314dccabfbe4dae09fea0