Update agdktunnel to use Cmake 3.22.1 by updating AGP
Currently a combination of building libraries from the command line with ./gradlew packageLocalZip -Plibraries=swappy -PpackageName=local and then building samples in Android Studio can result in failed builds with error messages referring to cmake versions.

Build tool/SDK versions are set by AGP version (which auto-downloads them on gradle build).

We should move to a single version of AGP in the whole repo; this is just a start.

Bug: 288242414
Test: N/A

Change-Id: I52cef55f625a7f9712d40f76a3f62eeab61a8ced