Modernize TuningForkMonitor app
See bug for justification.

API level 33 apps cannot be built with the Android Gradle Plugin specified (4.2.2). Upgrading the Android Gradle Plugin to 8.0 triggers a cascade of modernization requirements which this CL addresses.

Protobuf library was updated from 3.24.0 to 3.8.0. The Gradle configuration was updated to match the currently recommended format. The full (non-lite) version of the library was included which avoids the need to use simlinks to include the google/protobuf protos (a nice bonus is this makes the project easier to build on Windows where simlinks are unidiomatic). option=lite was set in the Gradle script to retain the use of lite accessors.

Kotlin was updated from 1.3.50 to 1.8.10. This requires changes to satisfy new restrictions around assigning const variables from non-final variables and requiring ‘when’ constructs to make a match in every case. In addition, Kotlin Synthetics were removed in 1.8, so their use to find views in layouts has been replaced with Jetpack View Binding.

AGP now requires that the launch activity is exported, but discourages the package from being specified there.

The structure of the build.gradle and settings.gradle files has been changed to the style recommended with the latest Android toolchain.

Bug: 302670554
Test: Manual
Change-Id: I5ec5864242b047cefda9abbf428f54846f5cad7f